backend/requirements.txt
Django
djangorestframework
djangorestframework-jwt
django-sslserver
pymysql
requests